requirements.txt
pyrogram
tgcrypto
httplib2
oauth2client
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
pymongo
flask